Message from Bugreport
The last missing point is, if the object is deleted first, we have to make sure, it is set to nil in property too. So this works:

  TIntfComp = class(TComponent)
    ...
  protected
    procedure Notification(AComponent: TComponent; Operation: TOperation); override;
    ...
  end;

...

procedure TIntfComp.Notification(AComponent: TComponent; Operation: TOperation);
begin
  inherited Notification(AComponent, Operation);
  // if ObjectHasInterface is being removed then we need to make sure it is
  // removed from here too
  if (Operation = opRemove) and Assigned(AComponent)
  and AComponent.IsImplementorOf(ObjectHasInterface) then
    Pointer(FObjectHasInterface) := nil;
end;

Now at least there should be problems with COM interfaces and TComponent, I can't see in my minimal examples here. So imho this bugreport can be closed. I update the package interface_test_pck.zip as a functional test project.